+++ src/server/serveractiveobject.cpp
@@ -17,7 +17,7 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 */
 
-#include "serverobject.h"
+#include "serveractiveobject.h"
 #include <fstream>
 #include "inventory.h"
 #include "constants.h" // BS
