+++ src/server/serveractiveobject.h
