+++ util/travis/clang-format-whitelist.txt
@@ -402,16 +402,14 @@ src/script/scripting_server.cpp
 src/script/scripting_server.h
 src/serialization.cpp
 src/serialization.h
-src/serveractiveobjectmap.cpp
-src/serveractiveobjectmap.h
 src/server.cpp
 src/serverenvironment.cpp
 src/serverenvironment.h
 src/server.h
 src/serverlist.cpp
 src/serverlist.h
-src/serverobject.cpp
-src/serverobject.h
+src/server/serveractiveobject.cpp
+src/server/serveractiveobject.h
 src/settings.cpp
 src/settings.h
 src/settings_translation_file.cpp
